Remove `red` from `comparison` defalt style
This is to avoid clashing with `rework` styling. It also avoids a
potential subconscious bias towards choosing blue over red (keep in mind
this is just my hunch; I never did actual testing on if this was the
case).

In addition, instead of explicitly creating the closures for every
color, this change moves to using a `map` call on an array of colors.